Update emblem and instructor display algorithms
This updates the emblem and instructor display algorithms to not only be more idiomatic in a React sense, but to also be quite a bit more performant. Also helps hide the empty emblem block that was taking up a lot of vertical space on mobile when displays intervals that had no emblems but still had an empty emblem div.